TP1/Functions/requirements.txt
functions-framework==3.*
SQLAlchemy==2.0.4
PyMySQL==1.0.3
gunicorn==20.1.0
cloud-sql-python-connector==1.2.1
google-cloud-datastore==2.10.0